backend/requirements.txt
Django==5.2.4
django-environ==0.12.0
psycopg[binary]==3.2.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djoser==2.3.3
django-cors-headers==4.7.0
django-modeltranslation==0.19.17
pillow==11.3.0
google-generativeai==0.8.5
google-auth==2.41.1
pytest==9.0.2
pytest-django==4.11.1
Faker==40.1.2
